+++ ImplicareCore/src/br/cefetmg/implicare/model/serviceImpl/TelefoneManagementImpl.java
@@ -20,62 +20,37 @@ public class TelefoneManagementImpl implements TelefoneManagement {
 
     @Override
     public boolean insert(Telefone Telefone) throws BusinessException, PersistenceException {
-        boolean result = false;
-        try {
-            result = TelefoneDao.insert(Telefone);
-            return result;
-        } catch (br.cefetmg.implicare.exception.PersistenceException ex) {
-            Logger.getLogger(TelefoneManagementImpl.class.getName()).log(Level.SEVERE, null, ex);
-        }
-        return result;
+        if(Telefone == null)
+            throw new BusinessException("Telefone não pode ser nulo");
+        return TelefoneDao.insert(Telefone);
+        
     }
 
     @Override
     public boolean update(Telefone Telefone) throws BusinessException, PersistenceException {
-        boolean result = false;
-        try {
-            result = TelefoneDao.update(Telefone);
-            return result;
-        } catch (br.cefetmg.implicare.exception.PersistenceException ex) {
-            Logger.getLogger(TelefoneManagementImpl.class.getName()).log(Level.SEVERE, null, ex);
-        }
-        return result;
+        if(Telefone == null)
+            throw new BusinessException("Telefone não pode ser nulo");
+        return TelefoneDao.update(Telefone);
     }
 
     @Override
     public boolean delete(Telefone Telefone) throws PersistenceException {
-        boolean result = false;
-        try {
-            result = TelefoneDao.delete(Telefone);
-            return result;
-        } catch (br.cefetmg.implicare.exception.PersistenceException ex) {
-            Logger.getLogger(TelefoneManagementImpl.class.getName()).log(Level.SEVERE, null, ex);
-        }
-        return result;
+        if(Telefone == null)
+            throw new BusinessException("Telefone não pode ser nulo");
+        return TelefoneDao.delete(Telefone);
+            
     }
 
     @Override
     public Telefone pesquisar(int Seq_Telefone) throws PersistenceException {
-        Telefone result = null;
-        try {
-            result = TelefoneDao.pesquisar(Seq_Telefone);
-            return result;
-        } catch (br.cefetmg.implicare.exception.PersistenceException ex) {
-            Logger.getLogger(TelefoneManagementImpl.class.getName()).log(Level.SEVERE, null, ex);
-        }
-        return result;
+        return TelefoneDao.pesquisar(Seq_Telefone);
+            
     }
 
     @Override
     public ArrayList<Telefone> listar(long CPF_CNPJ) throws PersistenceException {
-        ArrayList<Telefone> result = null;
-        try {
-            result = TelefoneDao.listar(CPF_CNPJ);
-            return result;
-        } catch (br.cefetmg.implicare.exception.PersistenceException ex) {
-            Logger.getLogger(TelefoneManagementImpl.class.getName()).log(Level.SEVERE, null, ex);
-        }
-        return result;
+        return TelefoneDao.listar(CPF_CNPJ);
+            
     }
 
 }
